add view model to login and register

+++ app/src/main/java/com/company/khomasi/domain/model/UserRegisterResponse.kt
@@ -27,9 +27,9 @@ data class UserDetails(
     @SerialName("lastName")
     val lastName: String,
     @SerialName("latitude")
-    val latitude: Int,
+    val latitude: Float,
     @SerialName("longitude")
-    val longitude: Int,
+    val longitude: Float,
     @SerialName("password")
     val password: String,
     @SerialName("phoneNumber")
